@@ -1,13 +1,32 @@
-from typing import Optional, Union
+import logging
+from typing import Any, Optional, Union
 
-from multiversx_sdk import LedgerAccount, Transaction, TransactionComputer
+from multiversx_sdk import Address, LedgerAccount, Transaction, TransactionComputer
+from multiversx_sdk.abi import (
+    AddressValue,
+    BigUIntValue,
+    BoolValue,
+    BytesValue,
+    StringValue,
+)
 
+from multiversx_sdk_cli.config_env import get_address_hrp
 from multiversx_sdk_cli.constants import (
+    ADDRESS_PREFIX,
     EXTRA_GAS_LIMIT_FOR_GUARDED_TRANSACTIONS,
     EXTRA_GAS_LIMIT_FOR_RELAYED_TRANSACTIONS,
+    FALSE_STR_LOWER,
+    HEX_PREFIX,
+    MAINCHAIN_ADDRESS_HRP,
+    STR_PREFIX,
+    TRUE_STR_LOWER,
 )
 from multiversx_sdk_cli.cosign_transaction import cosign_transaction
+from multiversx_sdk_cli.errors import BadUserInput, TransactionSigningError
+from multiversx_sdk_cli.guardian_relayer_data import GuardianRelayerData
 from multiversx_sdk_cli.interfaces import IAccount
+
+logger = logging.getLogger("base_controller")
 
 
 class BaseTransactionsController:
@@ -28,7 +47,10 @@ class BaseTransactionsController:
         self._set_options_for_hash_signing_if_needed(transaction, sender, guardian, relayer)
 
         if sender:
-            transaction.signature = sender.sign_transaction(transaction)
+            try:
+                transaction.signature = sender.sign_transaction(transaction)
+            except Exception as e:
+                raise TransactionSigningError(f"Could not sign transaction: {str(e)}")
 
         self._sign_guarded_transaction_if_guardian(
             transaction,
@@ -75,7 +97,10 @@ class BaseTransactionsController:
     ) -> Transaction:
         #  If the guardian account is provided, we sign locally. Otherwise, we reach for the trusted cosign service.
         if guardian:
-            transaction.guardian_signature = guardian.sign_transaction(transaction)
+            try:
+                transaction.guardian_signature = guardian.sign_transaction(transaction)
+            except Exception as e:
+                raise TransactionSigningError(f"Could not sign transaction: {str(e)}")
         elif transaction.guardian and guardian_service_url and guardian_2fa_code:
             cosign_transaction(transaction, guardian_service_url, guardian_2fa_code)
 
@@ -83,4 +108,65 @@ class BaseTransactionsController:
 
     def _sign_relayed_transaction_if_relayer(self, transaction: Transaction, relayer: Union[IAccount, None]):
         if relayer and transaction.relayer:
-            transaction.relayer_signature = relayer.sign_transaction(transaction)
+            try:
+                transaction.relayer_signature = relayer.sign_transaction(transaction)
+            except Exception as e:
+                raise TransactionSigningError(f"Could not sign transaction: {str(e)}")
+
+    def _convert_args_to_typed_values(self, arguments: list[str]) -> list[Any]:
+        args: list[Any] = []
+
+        for arg in arguments:
+            if arg.startswith(HEX_PREFIX):
+                args.append(BytesValue(self._hex_to_bytes(arg)))
+            elif arg.isnumeric():
+                args.append(BigUIntValue(int(arg)))
+            elif arg.startswith(ADDRESS_PREFIX):
+                args.append(AddressValue.new_from_address(Address.new_from_bech32(arg[len(ADDRESS_PREFIX) :])))
+            elif arg.startswith(MAINCHAIN_ADDRESS_HRP):
+                # this flow will be removed in the future
+                logger.warning(
+                    "Address argument has no prefix. This flow will be removed in the future. Please provide each address using the `addr:` prefix. (e.g. --arguments addr:erd1...)"
+                )
+                args.append(AddressValue.new_from_address(Address.new_from_bech32(arg)))
+            elif arg.startswith(get_address_hrp()):
+                args.append(AddressValue.new_from_address(Address.new_from_bech32(arg)))
+            elif arg.lower() == FALSE_STR_LOWER:
+                args.append(BoolValue(False))
+            elif arg.lower() == TRUE_STR_LOWER:
+                args.append(BoolValue(True))
+            elif arg.startswith(STR_PREFIX):
+                args.append(StringValue(arg[len(STR_PREFIX) :]))
+            else:
+                raise BadUserInput(
+                    f"Unknown argument type for argument: `{arg}`. Use `mxpy contract <sub-command> --help` to check all supported arguments"
+                )
+
+        return args
+
+    def _hex_to_bytes(self, arg: str):
+        argument = arg[len(HEX_PREFIX) :]
+        argument = argument.upper()
+        argument = self.ensure_even_length(argument)
+        return bytes.fromhex(argument)
+
+    def ensure_even_length(self, string: str) -> str:
+        if len(string) % 2 == 1:
+            return "0" + string
+        return string
+
+    def _set_transaction_fields(
+        self,
+        transaction: Transaction,
+        nonce: int,
+        version: int,
+        options: int,
+        gas_price: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ):
+        transaction.nonce = nonce
+        transaction.version = version
+        transaction.options = options
+        transaction.gas_price = gas_price
+        transaction.guardian = guardian_and_relayer_data.guardian_address
+        transaction.relayer = guardian_and_relayer_data.relayer_address
