@@ -10,20 +10,28 @@ from multiversx_sdk import LibraryConfig
 from rich.logging import RichHandler
 
 import multiversx_sdk_cli.cli_config
+import multiversx_sdk_cli.cli_config_env
+import multiversx_sdk_cli.cli_config_wallet
 import multiversx_sdk_cli.cli_contracts
 import multiversx_sdk_cli.cli_data
 import multiversx_sdk_cli.cli_delegation
 import multiversx_sdk_cli.cli_deps
 import multiversx_sdk_cli.cli_dns
 import multiversx_sdk_cli.cli_faucet
+import multiversx_sdk_cli.cli_get
+import multiversx_sdk_cli.cli_governance
 import multiversx_sdk_cli.cli_ledger
 import multiversx_sdk_cli.cli_localnet
+import multiversx_sdk_cli.cli_multisig
 import multiversx_sdk_cli.cli_transactions
 import multiversx_sdk_cli.cli_validator_wallet
 import multiversx_sdk_cli.cli_validators
 import multiversx_sdk_cli.cli_wallet
 import multiversx_sdk_cli.version
 from multiversx_sdk_cli import config, errors, utils, ux
+from multiversx_sdk_cli.cli_shared import set_proxy_from_config_if_not_provided
+from multiversx_sdk_cli.config_env import get_address_hrp
+from multiversx_sdk_cli.constants import LOG_LEVELS, SDK_PATH
 
 logger = logging.getLogger("cli")
 
@@ -42,11 +50,12 @@ def main(cli_args: list[str] = sys.argv[1:]):
 
 
 def _do_main(cli_args: list[str]):
-    utils.ensure_folder(config.SDK_PATH)
-    argv_with_config_args = config.add_config_args(cli_args)
-    parser = setup_parser(argv_with_config_args)
+    utils.ensure_folder(SDK_PATH)
+    parser = setup_parser(cli_args)
     argcomplete.autocomplete(parser)
-    args = parser.parse_args(argv_with_config_args)
+
+    _handle_global_arguments(cli_args)
+    args = parser.parse_args(cli_args)
 
     if args.verbose:
         logging.basicConfig(
@@ -56,22 +65,21 @@ def _do_main(cli_args: list[str]):
             handlers=[RichHandler(show_time=False, rich_tracebacks=True)],
         )
     else:
+        level: str = args.log_level
         logging.basicConfig(
-            level="INFO",
+            level=level.upper(),
             format="%(name)s: %(message)s",
             handlers=[RichHandler(show_time=False, rich_tracebacks=True)],
         )
 
     verify_deprecated_entries_in_config_file()
-    default_hrp = config.get_address_hrp()
+    default_hrp = get_address_hrp()
     LibraryConfig.default_address_hrp = default_hrp
-
-    if hasattr(args, "recall_nonce") and args.recall_nonce:
-        logger.warning("The --recall-nonce flag is DEPRECATED. The nonce is fetched from the network by default.")
 
     if not hasattr(args, "func"):
         parser.print_help()
     else:
+        set_proxy_from_config_if_not_provided(args)
         args.func(args)
 
 
@@ -105,10 +113,18 @@ See:
         version=f"MultiversX Python CLI (mxpy) {version}",
     )
     parser.add_argument("--verbose", action="store_true", default=False)
+    parser.add_argument(
+        "--log-level",
+        type=str,
+        default=config.get_log_level_from_config(),
+        choices=LOG_LEVELS,
+        help="default: %(default)s",
+    )
 
     subparsers = parser.add_subparsers()
     commands: list[Any] = []
 
+    commands.append(multiversx_sdk_cli.cli_config_wallet.setup_parser(subparsers))
     commands.append(multiversx_sdk_cli.cli_contracts.setup_parser(args, subparsers))
     commands.append(multiversx_sdk_cli.cli_transactions.setup_parser(args, subparsers))
     commands.append(multiversx_sdk_cli.cli_validators.setup_parser(args, subparsers))
@@ -122,6 +138,10 @@ See:
     commands.append(multiversx_sdk_cli.cli_delegation.setup_parser(args, subparsers))
     commands.append(multiversx_sdk_cli.cli_dns.setup_parser(args, subparsers))
     commands.append(multiversx_sdk_cli.cli_faucet.setup_parser(args, subparsers))
+    commands.append(multiversx_sdk_cli.cli_multisig.setup_parser(args, subparsers))
+    commands.append(multiversx_sdk_cli.cli_governance.setup_parser(args, subparsers))
+    commands.append(multiversx_sdk_cli.cli_config_env.setup_parser(subparsers))
+    commands.append(multiversx_sdk_cli.cli_get.setup_parser(subparsers))
 
     parser.epilog = """
 ----------------------
@@ -145,6 +165,26 @@ def verify_deprecated_entries_in_config_file():
         message += f"-> {entry} \n"
 
     ux.show_warning(message.rstrip("\n"))
+
+
+def _handle_global_arguments(args: list[str]):
+    """
+    Handle global arguments like --verbose and --log-level.
+    """
+    log_level_arg = "--log-level"
+    if log_level_arg in args:
+        index = args.index(log_level_arg)
+        if index + 1 >= len(args):
+            raise ValueError(f"Argument {log_level_arg} must be followed by a log level value.")
+
+        log_arg = args.pop(index)
+        log_value = args.pop(index)
+        args.insert(0, log_value)
+        args.insert(0, log_arg)
+
+    if "--verbose" in args:
+        args.remove("--verbose")
+        args.insert(0, "--verbose")
 
 
 if __name__ == "__main__":
