@@ -2,19 +2,11 @@ import logging
 from pathlib import Path
 from typing import Any
 
-from multiversx_sdk import (
-    Address,
-    ProxyNetworkProvider,
-    Token,
-    TokenComputer,
-    TokenTransfer,
-    TransactionComputer,
-)
+from multiversx_sdk import Address, ProxyNetworkProvider, TransactionComputer
 
 from multiversx_sdk_cli import cli_shared, utils
 from multiversx_sdk_cli.args_validation import (
     ensure_relayer_wallet_args_are_provided,
-    ensure_wallet_args_are_provided,
     validate_broadcast_args,
     validate_chain_id_args,
     validate_nonce_args,
@@ -51,17 +43,7 @@ def setup_parser(args: list[str], subparsers: Any) -> Any:
     cli_shared.add_guardian_wallet_args(args, sub)
     cli_shared.add_relayed_v3_wallet_args(args, sub)
 
-    sub.add_argument(
-        "--wait-result",
-        action="store_true",
-        default=False,
-        help="signal to wait for the transaction result - only valid if --send is set",
-    )
-    sub.add_argument(
-        "--timeout",
-        default=100,
-        help="max num of seconds to wait for result" " - only valid if --wait-result is set",
-    )
+    cli_shared.add_wait_result_and_timeout_args(sub)
     sub.set_defaults(func=create_transaction)
 
     sub = cli_shared.add_command_subparser(
@@ -116,7 +98,6 @@ def _add_common_arguments(args: list[str], sub: Any):
 def create_transaction(args: Any):
     validate_nonce_args(args)
     validate_receiver_args(args)
-    ensure_wallet_args_are_provided(args)
     validate_broadcast_args(args)
     validate_chain_id_args(args)
 
@@ -133,9 +114,9 @@ def create_transaction(args: Any):
     gas_limit = int(args.gas_limit) if args.gas_limit else 0
 
     transfers = getattr(args, "token_transfers", None)
-    transfers = prepare_token_transfers(transfers) if transfers else None
+    transfers = cli_shared.prepare_token_transfers(transfers) if transfers else None
 
-    chain_id = cli_shared.get_chain_id(args.chain, args.proxy)
+    chain_id = cli_shared.get_chain_id(args.proxy, args.chain)
     tx_controller = TransactionsController(chain_id)
 
     tx = tx_controller.create_transaction(
@@ -155,22 +136,6 @@ def create_transaction(args: Any):
     cli_shared.send_or_simulate(tx, args)
 
 
-def prepare_token_transfers(transfers: list[Any]) -> list[TokenTransfer]:
-    token_computer = TokenComputer()
-    token_transfers: list[TokenTransfer] = []
-
-    for i in range(0, len(transfers) - 1, 2):
-        identifier = transfers[i]
-        amount = int(transfers[i + 1])
-        nonce = token_computer.extract_nonce_from_extended_identifier(identifier)
-
-        token = Token(identifier, nonce)
-        transfer = TokenTransfer(token, amount)
-        token_transfers.append(transfer)
-
-    return token_transfers
-
-
 def send_transaction(args: Any):
     validate_proxy_argument(args)
 
@@ -181,6 +146,8 @@ def send_transaction(args: Any):
     proxy = ProxyNetworkProvider(url=args.proxy, config=config)
 
     try:
+        cli_shared._confirm_continuation_if_required(tx)
+
         tx_hash = proxy.send_transaction(tx)
         output.set_emitted_transaction_hash(tx_hash.hex())
     finally:
@@ -195,7 +162,8 @@ def sign_transaction(args: Any):
 
     try:
         sender = cli_shared.prepare_account(args)
-    except NoWalletProvided:
+    except:
+        logger.info("No sender wallet provided. Will not sign for the sender.")
         sender = None
 
     if sender and sender.address != tx.sender:
