@@ -5,7 +5,6 @@ from multiversx_sdk import Address, ValidatorPublicKey, ValidatorsSigners
 
 from multiversx_sdk_cli import cli_shared, utils
 from multiversx_sdk_cli.args_validation import (
-    ensure_wallet_args_are_provided,
     validate_broadcast_args,
     validate_chain_id_args,
     validate_nonce_args,
@@ -148,7 +147,6 @@ def _add_nodes_arg(sub: Any):
 def validate_args(args: Any) -> None:
     validate_nonce_args(args)
     validate_receiver_args(args)
-    ensure_wallet_args_are_provided(args)
     validate_broadcast_args(args)
     validate_chain_id_args(args)
 
@@ -197,7 +195,7 @@ def do_stake(args: Any):
 
 
 def _get_validators_controller(args: Any):
-    chain_id = cli_shared.get_chain_id(args.chain, args.proxy)
+    chain_id = cli_shared.get_chain_id(args.proxy, args.chain)
     validators = ValidatorsController(chain_id)
     return validators
 
@@ -219,6 +217,7 @@ def _parse_public_bls_keys(public_bls_keys: str) -> list[ValidatorPublicKey]:
 
 def do_unstake(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -247,6 +246,7 @@ def do_unstake(args: Any):
 
 def do_unjail(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -275,6 +275,7 @@ def do_unjail(args: Any):
 
 def do_unbond(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -303,6 +304,7 @@ def do_unbond(args: Any):
 
 def change_reward_address(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -331,6 +333,7 @@ def change_reward_address(args: Any):
 
 def do_claim(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -357,6 +360,7 @@ def do_claim(args: Any):
 
 def do_unstake_nodes(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -386,6 +390,7 @@ def do_unstake_nodes(args: Any):
 
 def do_unstake_tokens(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -414,6 +419,7 @@ def do_unstake_tokens(args: Any):
 
 def do_unbond_nodes(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -443,6 +449,7 @@ def do_unbond_nodes(args: Any):
 
 def do_unbond_tokens(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -471,6 +478,7 @@ def do_unbond_tokens(args: Any):
 
 def do_clean_registered_data(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
@@ -497,6 +505,7 @@ def do_clean_registered_data(args: Any):
 
 def do_restake_unstaked_nodes(args: Any):
     validate_args(args)
+
     sender = cli_shared.prepare_sender(args)
     guardian_and_relayer_data = cli_shared.get_guardian_and_relayer_data(
         sender=sender.address.to_bech32(),
