@@ -1,31 +1,16 @@
 import os
+from functools import cache
 from pathlib import Path
 from typing import Any
 
 from multiversx_sdk import NetworkProviderConfig
 
 from multiversx_sdk_cli import errors, utils
+from multiversx_sdk_cli.constants import LOG_LEVELS, SDK_PATH
+from multiversx_sdk_cli.ux import show_warning
 
-SDK_PATH = Path("~/multiversx-sdk").expanduser().resolve()
 LOCAL_CONFIG_PATH = Path("mxpy.json").resolve()
 GLOBAL_CONFIG_PATH = SDK_PATH / "mxpy.json"
-
-
-class MetaChainSystemSCsCost:
-    STAKE = 5000000
-    UNSTAKE = 5000000
-    UNBOND = 5000000
-    CLAIM = 5000000
-    GET = 5000000
-    CHANGE_REWARD_ADDRESS = 5000000
-    CHANGE_VALIDATOR_KEYS = 5000000
-    UNJAIL = 5000000
-    DELEGATION_MANAGER_OPS = 50000000
-    DELEGATION_OPS = 1000000
-    UNSTAKE_TOKENS = 5000000
-    UNBOND_TOKENS = 5000000
-    CLEAN_REGISTERED_DATA = 5000000
-    RESTAKE_UNSTAKED_NODES = 5000000
 
 
 def get_dependency_resolution(key: str) -> str:
@@ -48,6 +33,7 @@ def get_dependency_url(key: str, tag: str, platform: str) -> str:
     return url_template.replace("{TAG}", tag)
 
 
+@cache
 def get_value(name: str) -> str:
     _guard_valid_name(name)
     data = get_active()
@@ -55,10 +41,6 @@ def get_value(name: str) -> str:
     value = data.get(name, default_value)
     assert isinstance(value, str)
     return value
-
-
-def get_address_hrp() -> str:
-    return get_value("default_address_hrp")
 
 
 def set_value(name: str, value: Any):
@@ -145,26 +127,28 @@ def _guard_valid_config_deletion(name: str):
 
 def get_defaults() -> dict[str, Any]:
     return {
-        "dependencies.vmtools.tag": "v1.5.24",
-        "dependencies.vmtools.urlTemplate.linux": "https://github.com/multiversx/mx-chain-vm-go/archive/{TAG}.tar.gz",
-        "dependencies.vmtools.urlTemplate.osx": "https://github.com/multiversx/mx-chain-vm-go/archive/{TAG}.tar.gz",
-        "dependencies.vmtools.urlTemplate.windows": "https://github.com/multiversx/mx-chain-vm-go/archive/{TAG}.tar.gz",
-        "dependencies.rust.tag": "stable",
         "dependencies.golang.resolution": "SDK",
         "dependencies.golang.tag": "go1.20.7",
         "dependencies.golang.urlTemplate.linux": "https://golang.org/dl/{TAG}.linux-amd64.tar.gz",
         "dependencies.golang.urlTemplate.osx": "https://golang.org/dl/{TAG}.darwin-amd64.tar.gz",
         "dependencies.golang.urlTemplate.windows": "https://golang.org/dl/{TAG}.windows-amd64.zip",
-        "dependencies.twiggy.tag": "",
-        "dependencies.sc-meta.tag": "",
         "dependencies.testwallets.tag": "v1.0.0",
         "dependencies.testwallets.urlTemplate.linux": "https://github.com/multiversx/mx-sdk-testwallets/archive/{TAG}.tar.gz",
         "dependencies.testwallets.urlTemplate.osx": "https://github.com/multiversx/mx-sdk-testwallets/archive/{TAG}.tar.gz",
         "dependencies.testwallets.urlTemplate.windows": "https://github.com/multiversx/mx-sdk-testwallets/archive/{TAG}.tar.gz",
-        "dependencies.wasm-opt.tag": "0.112.0",
         "github_api_token": "",
-        "default_address_hrp": "erd",
+        "log_level": "info",
     }
+
+
+def get_log_level_from_config() -> str:
+    log_level = get_value("log_level")
+    if log_level not in LOG_LEVELS:
+        default_log_level: str = get_defaults()["log_level"]
+        show_warning(f"Invalid log level set in config: [{log_level}]. Defaulting to [{default_log_level}].")
+        return default_log_level
+
+    return log_level
 
 
 def get_deprecated_entries_in_config_file():
@@ -179,6 +163,7 @@ def resolve_config_path() -> Path:
     return GLOBAL_CONFIG_PATH
 
 
+@cache
 def read_file() -> dict[str, Any]:
     config_path = resolve_config_path()
     if config_path.exists():
@@ -190,53 +175,6 @@ def read_file() -> dict[str, Any]:
 def write_file(data: dict[str, Any]):
     config_path = resolve_config_path()
     utils.write_json_file(str(config_path), data)
-
-
-def add_config_args(argv: list[str]) -> list[str]:
-    try:
-        command, subcommand, *_ = argv
-    except ValueError:
-        return argv
-
-    config = read_file()
-
-    try:
-        config_args = config[command][subcommand]
-    except KeyError:
-        return argv
-
-    final_args = determine_final_args(argv, config_args)
-    print(f"Found extra arguments in mxpy.json. Final arguments: {final_args}")
-    return final_args
-
-
-def determine_final_args(argv: list[str], config_args: dict[str, Any]) -> list[str]:
-    extra_args: list[str] = []
-    for key, value in config_args.items():
-        key_arg = f"--{key}"
-        # arguments from the command line override the config
-        if key_arg in argv:
-            continue
-        if any(arg.startswith(f"{key_arg}=") for arg in argv):
-            continue
-        extra_args.append(key_arg)
-        if value is True:
-            continue
-        if isinstance(value, list):
-            for item in value:  # type: ignore
-                extra_args.append(str(item))  # type: ignore
-        else:
-            extra_args.append(str(value))
-
-    # the verbose flag is an exception since it has to go before the command and subcommand
-    # eg. mxpy --verbose contract deploy
-    verbose_flag = "--verbose"
-    pre_args = []
-    if verbose_flag in extra_args:
-        extra_args.remove(verbose_flag)
-        pre_args = [verbose_flag]
-
-    return pre_args + argv + extra_args
 
 
 def get_dependency_directory(key: str, tag: str) -> Path:
