@@ -10,12 +10,11 @@ from multiversx_sdk import (
 
 from multiversx_sdk_cli import cli_shared
 from multiversx_sdk_cli.args_validation import (
-    ensure_wallet_args_are_provided,
     validate_broadcast_args,
     validate_chain_id_args,
     validate_transaction_args,
 )
-from multiversx_sdk_cli.config import get_address_hrp
+from multiversx_sdk_cli.config_env import get_address_hrp
 from multiversx_sdk_cli.constants import ADDRESS_ZERO_HEX
 from multiversx_sdk_cli.transactions import TransactionsController
 
@@ -62,7 +61,6 @@ def validate_name(name: str, shard_id: int, proxy: INetworkProvider):
 
 def register(args: Any):
     validate_transaction_args(args)
-    ensure_wallet_args_are_provided(args)
     validate_broadcast_args(args)
     validate_chain_id_args(args)
 
@@ -77,7 +75,7 @@ def register(args: Any):
     receiver = dns_address_for_name(args.name)
     data = dns_register_data(args.name)
 
-    chain_id = cli_shared.get_chain_id(args.chain, args.proxy)
+    chain_id = cli_shared.get_chain_id(args.proxy, args.chain)
     controller = TransactionsController(chain_id)
 
     tx = controller.create_transaction(
