@@ -1,55 +1,61 @@
+import logging
+from pathlib import Path
+from typing import Any, Optional
+
 from multiversx_sdk import (
     Address,
-    DelegationTransactionsFactory,
+    MultisigTransactionsFactory,
+    TokenTransfer,
     Transaction,
     TransactionsFactoryConfig,
-    ValidatorPublicKey,
 )
-from multiversx_sdk.abi import BigUIntValue, Serializer
+from multiversx_sdk.abi import Abi
 
 from multiversx_sdk_cli.base_transactions_controller import BaseTransactionsController
-from multiversx_sdk_cli.config_env import get_address_hrp
-from multiversx_sdk_cli.errors import BadUsage
 from multiversx_sdk_cli.guardian_relayer_data import GuardianRelayerData
 from multiversx_sdk_cli.interfaces import IAccount
 
-DELEGATION_MANAGER_SC_ADDRESS_HEX = "000000000000000000010000000000000000000000000000000000000004ffff"
+logger = logging.getLogger("multisig")
 
 
-class DelegationOperations(BaseTransactionsController):
-    def __init__(self, config: TransactionsFactoryConfig) -> None:
-        self._factory = DelegationTransactionsFactory(config)
+class MultisigWrapper(BaseTransactionsController):
+    def __init__(self, config: TransactionsFactoryConfig, abi: Abi):
+        self._factory = MultisigTransactionsFactory(config, abi)
 
-    def prepare_transaction_for_new_delegation_contract(
+    def prepare_deploy_transaction(
         self,
         owner: IAccount,
-        native_amount: int,
-        total_delegation_cap: int,
-        service_fee: int,
+        nonce: int,
+        bytecode: Path,
+        quorum: int,
+        board_members: list[Address],
+        upgradeable: bool,
+        readable: bool,
+        payable: bool,
+        payable_by_sc: bool,
         gas_limit: int,
         gas_price: int,
-        nonce: int,
         version: int,
         options: int,
         guardian_and_relayer_data: GuardianRelayerData,
     ) -> Transaction:
-        tx = self._factory.create_transaction_for_new_delegation_contract(
+        tx = self._factory.create_transaction_for_deploy(
             sender=owner.address,
-            total_delegation_cap=total_delegation_cap,
-            service_fee=service_fee,
-            amount=native_amount,
+            bytecode=bytecode,
+            quorum=quorum,
+            board=board_members,
+            gas_limit=gas_limit,
+            is_upgradeable=upgradeable,
+            is_readable=readable,
+            is_payable=payable,
+            is_payable_by_sc=payable_by_sc,
         )
-        tx.gas_price = gas_price
         tx.nonce = nonce
         tx.version = version
         tx.options = options
+        tx.gas_price = gas_price
         tx.guardian = guardian_and_relayer_data.guardian_address
         tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
 
         self.sign_transaction(
             transaction=tx,
@@ -62,38 +68,32 @@ class DelegationOperations(BaseTransactionsController):
 
         return tx
 
-    def prepare_transaction_for_adding_nodes(
+    def prepare_deposit_transaction(
         self,
         owner: IAccount,
-        delegation_contract: Address,
-        public_keys: list[ValidatorPublicKey],
-        signed_messages: list[bytes],
+        nonce: int,
+        contract: Address,
         gas_limit: int,
         gas_price: int,
-        value: int,
-        nonce: int,
         version: int,
         options: int,
         guardian_and_relayer_data: GuardianRelayerData,
+        native_amount: Optional[int] = None,
+        token_transfers: Optional[list[TokenTransfer]] = None,
     ) -> Transaction:
-        tx = self._factory.create_transaction_for_adding_nodes(
+        tx = self._factory.create_transaction_for_deposit(
             sender=owner.address,
-            delegation_contract=delegation_contract,
-            public_keys=public_keys,
-            signed_messages=signed_messages,
+            contract=contract,
+            gas_limit=gas_limit,
+            native_token_amount=native_amount,
+            token_transfers=token_transfers,
         )
-        tx.value = value
-        tx.gas_price = gas_price
         tx.nonce = nonce
         tx.version = version
         tx.options = options
+        tx.gas_price = gas_price
         tx.guardian = guardian_and_relayer_data.guardian_address
         tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
 
         self.sign_transaction(
             transaction=tx,
@@ -106,36 +106,30 @@ class DelegationOperations(BaseTransactionsController):
 
         return tx
 
-    def prepare_transaction_for_removing_nodes(
+    def prepare_discard_action_transaction(
         self,
         owner: IAccount,
-        delegation_contract: Address,
-        public_keys: list[ValidatorPublicKey],
+        nonce: int,
+        contract: Address,
+        action_id: int,
         gas_limit: int,
         gas_price: int,
-        value: int,
-        nonce: int,
         version: int,
         options: int,
         guardian_and_relayer_data: GuardianRelayerData,
     ) -> Transaction:
-        tx = self._factory.create_transaction_for_removing_nodes(
+        tx = self._factory.create_transaction_for_discard_action(
             sender=owner.address,
-            delegation_contract=delegation_contract,
-            public_keys=public_keys,
+            contract=contract,
+            action_id=action_id,
+            gas_limit=gas_limit,
         )
-        tx.value = value
-        tx.gas_price = gas_price
         tx.nonce = nonce
         tx.version = version
         tx.options = options
+        tx.gas_price = gas_price
         tx.guardian = guardian_and_relayer_data.guardian_address
         tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
 
         self.sign_transaction(
             transaction=tx,
@@ -148,36 +142,30 @@ class DelegationOperations(BaseTransactionsController):
 
         return tx
 
-    def prepare_transaction_for_staking_nodes(
+    def prepare_discard_batch_transaction(
         self,
         owner: IAccount,
-        delegation_contract: Address,
-        public_keys: list[ValidatorPublicKey],
+        nonce: int,
+        contract: Address,
+        action_ids: list[int],
         gas_limit: int,
         gas_price: int,
-        value: int,
-        nonce: int,
         version: int,
         options: int,
         guardian_and_relayer_data: GuardianRelayerData,
     ) -> Transaction:
-        tx = self._factory.create_transaction_for_staking_nodes(
+        tx = self._factory.create_transaction_for_discard_batch(
             sender=owner.address,
-            delegation_contract=delegation_contract,
-            public_keys=public_keys,
+            contract=contract,
+            action_ids=action_ids,
+            gas_limit=gas_limit,
         )
-        tx.value = value
-        tx.gas_price = gas_price
         tx.nonce = nonce
         tx.version = version
         tx.options = options
+        tx.gas_price = gas_price
         tx.guardian = guardian_and_relayer_data.guardian_address
         tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
 
         self.sign_transaction(
             transaction=tx,
@@ -190,36 +178,30 @@ class DelegationOperations(BaseTransactionsController):
 
         return tx
 
-    def prepare_transaction_for_unbonding_nodes(
+    def prepare_add_board_member_transaction(
         self,
         owner: IAccount,
-        delegation_contract: Address,
-        public_keys: list[ValidatorPublicKey],
+        nonce: int,
+        contract: Address,
+        board_member: Address,
         gas_limit: int,
         gas_price: int,
-        value: int,
-        nonce: int,
         version: int,
         options: int,
         guardian_and_relayer_data: GuardianRelayerData,
     ) -> Transaction:
-        tx = self._factory.create_transaction_for_unbonding_nodes(
+        tx = self._factory.create_transaction_for_propose_add_board_member(
             sender=owner.address,
-            delegation_contract=delegation_contract,
-            public_keys=public_keys,
+            contract=contract,
+            board_member=board_member,
+            gas_limit=gas_limit,
         )
-        tx.value = value
-        tx.gas_price = gas_price
         tx.nonce = nonce
         tx.version = version
         tx.options = options
+        tx.gas_price = gas_price
         tx.guardian = guardian_and_relayer_data.guardian_address
         tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
 
         self.sign_transaction(
             transaction=tx,
@@ -232,36 +214,30 @@ class DelegationOperations(BaseTransactionsController):
 
         return tx
 
-    def prepare_transaction_for_unstaking_nodes(
+    def prepare_add_proposer_transaction(
         self,
         owner: IAccount,
-        delegation_contract: Address,
-        public_keys: list[ValidatorPublicKey],
+        nonce: int,
+        contract: Address,
+        proposer: Address,
         gas_limit: int,
         gas_price: int,
-        value: int,
-        nonce: int,
         version: int,
         options: int,
         guardian_and_relayer_data: GuardianRelayerData,
     ) -> Transaction:
-        tx = self._factory.create_transaction_for_unstaking_nodes(
+        tx = self._factory.create_transaction_for_propose_add_proposer(
             sender=owner.address,
-            delegation_contract=delegation_contract,
-            public_keys=public_keys,
+            contract=contract,
+            proposer=proposer,
+            gas_limit=gas_limit,
         )
-        tx.value = value
-        tx.gas_price = gas_price
         tx.nonce = nonce
         tx.version = version
         tx.options = options
+        tx.gas_price = gas_price
         tx.guardian = guardian_and_relayer_data.guardian_address
         tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
 
         self.sign_transaction(
             transaction=tx,
@@ -274,36 +250,30 @@ class DelegationOperations(BaseTransactionsController):
 
         return tx
 
-    def prepare_transaction_for_unjailing_nodes(
+    def prepare_remove_user_transaction(
         self,
         owner: IAccount,
-        delegation_contract: Address,
-        public_keys: list[ValidatorPublicKey],
+        nonce: int,
+        contract: Address,
+        user: Address,
         gas_limit: int,
         gas_price: int,
-        value: int,
-        nonce: int,
         version: int,
         options: int,
         guardian_and_relayer_data: GuardianRelayerData,
     ) -> Transaction:
-        tx = self._factory.create_transaction_for_unjailing_nodes(
+        tx = self._factory.create_transaction_for_propose_remove_user(
             sender=owner.address,
-            delegation_contract=delegation_contract,
-            public_keys=public_keys,
-            amount=value,
+            contract=contract,
+            user=user,
+            gas_limit=gas_limit,
         )
-        tx.gas_price = gas_price
         tx.nonce = nonce
         tx.version = version
         tx.options = options
+        tx.gas_price = gas_price
         tx.guardian = guardian_and_relayer_data.guardian_address
         tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
 
         self.sign_transaction(
             transaction=tx,
@@ -316,34 +286,30 @@ class DelegationOperations(BaseTransactionsController):
 
         return tx
 
-    def prepare_transaction_for_delegating(
+    def prepare_change_quorum_transaction(
         self,
         owner: IAccount,
-        delegation_contract: Address,
+        nonce: int,
+        contract: Address,
+        quorum: int,
         gas_limit: int,
         gas_price: int,
-        value: int,
-        nonce: int,
         version: int,
         options: int,
         guardian_and_relayer_data: GuardianRelayerData,
     ) -> Transaction:
-        tx = self._factory.create_transaction_for_delegating(
+        tx = self._factory.create_transaction_for_propose_change_quorum(
             sender=owner.address,
-            delegation_contract=delegation_contract,
-            amount=value,
+            contract=contract,
+            quorum=quorum,
+            gas_limit=gas_limit,
         )
-        tx.gas_price = gas_price
         tx.nonce = nonce
         tx.version = version
         tx.options = options
+        tx.gas_price = gas_price
         tx.guardian = guardian_and_relayer_data.guardian_address
         tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
 
         self.sign_transaction(
             transaction=tx,
@@ -356,435 +322,587 @@ class DelegationOperations(BaseTransactionsController):
 
         return tx
 
-    def prepare_transaction_for_claiming_rewards(
+    def prepare_transfer_execute_transaction(
         self,
         owner: IAccount,
-        delegation_contract: Address,
+        nonce: int,
+        contract: Address,
+        receiver: Address,
+        native_token_amount: int,
         gas_limit: int,
         gas_price: int,
-        value: int,
-        nonce: int,
         version: int,
         options: int,
+        should_prepare_args_for_factory: bool,
         guardian_and_relayer_data: GuardianRelayerData,
+        opt_gas_limit: Optional[int] = None,
+        abi: Optional[Abi] = None,
+        function: Optional[str] = None,
+        arguments: Optional[list[Any]] = None,
     ) -> Transaction:
-        tx = self._factory.create_transaction_for_claiming_rewards(
-            sender=owner.address, delegation_contract=delegation_contract
-        )
-        tx.value = value
-        tx.gas_price = gas_price
-        tx.nonce = nonce
-        tx.version = version
-        tx.options = options
-        tx.guardian = guardian_and_relayer_data.guardian_address
-        tx.relayer = guardian_and_relayer_data.relayer_address
+        args = arguments if arguments else []
+        if should_prepare_args_for_factory:
+            args = self._convert_args_to_typed_values(args)
 
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
-
-        self.sign_transaction(
-            transaction=tx,
-            sender=owner,
-            guardian=guardian_and_relayer_data.guardian,
-            relayer=guardian_and_relayer_data.relayer,
-            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
-            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
-        )
-
-        return tx
-
-    def prepare_transaction_for_redelegating_rewards(
-        self,
-        owner: IAccount,
-        delegation_contract: Address,
-        gas_limit: int,
-        gas_price: int,
-        value: int,
-        nonce: int,
-        version: int,
-        options: int,
-        guardian_and_relayer_data: GuardianRelayerData,
-    ) -> Transaction:
-        tx = self._factory.create_transaction_for_redelegating_rewards(
-            sender=owner.address, delegation_contract=delegation_contract
-        )
-        tx.value = value
-        tx.gas_price = gas_price
-        tx.nonce = nonce
-        tx.version = version
-        tx.options = options
-        tx.guardian = guardian_and_relayer_data.guardian_address
-        tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
-
-        self.sign_transaction(
-            transaction=tx,
-            sender=owner,
-            guardian=guardian_and_relayer_data.guardian,
-            relayer=guardian_and_relayer_data.relayer,
-            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
-            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
-        )
-
-        return tx
-
-    def prepare_transaction_for_undelegating(
-        self,
-        owner: IAccount,
-        delegation_contract: Address,
-        gas_limit: int,
-        gas_price: int,
-        value: int,
-        nonce: int,
-        version: int,
-        options: int,
-        guardian_and_relayer_data: GuardianRelayerData,
-    ) -> Transaction:
-        tx = self._factory.create_transaction_for_undelegating(
+        tx = self._factory.create_transaction_for_propose_transfer_execute(
             sender=owner.address,
-            delegation_contract=delegation_contract,
-            amount=value,
-        )
-        tx.gas_price = gas_price
-        tx.nonce = nonce
-        tx.version = version
-        tx.options = options
-        tx.guardian = guardian_and_relayer_data.guardian_address
-        tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
-
-        self.sign_transaction(
-            transaction=tx,
-            sender=owner,
-            guardian=guardian_and_relayer_data.guardian,
-            relayer=guardian_and_relayer_data.relayer,
-            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
-            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
-        )
-
-        return tx
-
-    def prepare_transaction_for_withdrawing(
-        self,
-        owner: IAccount,
-        delegation_contract: Address,
-        gas_limit: int,
-        gas_price: int,
-        value: int,
-        nonce: int,
-        version: int,
-        options: int,
-        guardian_and_relayer_data: GuardianRelayerData,
-    ) -> Transaction:
-        tx = self._factory.create_transaction_for_withdrawing(
-            sender=owner.address, delegation_contract=delegation_contract
-        )
-        tx.value = value
-        tx.gas_price = gas_price
-        tx.nonce = nonce
-        tx.version = version
-        tx.options = options
-        tx.guardian = guardian_and_relayer_data.guardian_address
-        tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
-
-        self.sign_transaction(
-            transaction=tx,
-            sender=owner,
-            guardian=guardian_and_relayer_data.guardian,
-            relayer=guardian_and_relayer_data.relayer,
-            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
-            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
-        )
-
-        return tx
-
-    def prepare_transaction_for_changing_service_fee(
-        self,
-        owner: IAccount,
-        delegation_contract: Address,
-        service_fee: int,
-        gas_limit: int,
-        gas_price: int,
-        value: int,
-        nonce: int,
-        version: int,
-        options: int,
-        guardian_and_relayer_data: GuardianRelayerData,
-    ) -> Transaction:
-        tx = self._factory.create_transaction_for_changing_service_fee(
-            sender=owner.address,
-            delegation_contract=delegation_contract,
-            service_fee=service_fee,
-        )
-        tx.value = value
-        tx.gas_price = gas_price
-        tx.nonce = nonce
-        tx.version = version
-        tx.options = options
-        tx.guardian = guardian_and_relayer_data.guardian_address
-        tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
-
-        self.sign_transaction(
-            transaction=tx,
-            sender=owner,
-            guardian=guardian_and_relayer_data.guardian,
-            relayer=guardian_and_relayer_data.relayer,
-            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
-            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
-        )
-
-        return tx
-
-    def prepare_transaction_for_modifying_delegation_cap(
-        self,
-        owner: IAccount,
-        delegation_contract: Address,
-        delegation_cap: int,
-        gas_limit: int,
-        gas_price: int,
-        value: int,
-        nonce: int,
-        version: int,
-        options: int,
-        guardian_and_relayer_data: GuardianRelayerData,
-    ) -> Transaction:
-        tx = self._factory.create_transaction_for_modifying_delegation_cap(
-            sender=owner.address,
-            delegation_contract=delegation_contract,
-            delegation_cap=delegation_cap,
-        )
-        tx.value = value
-        tx.gas_price = gas_price
-        tx.nonce = nonce
-        tx.version = version
-        tx.options = options
-        tx.guardian = guardian_and_relayer_data.guardian_address
-        tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
-
-        self.sign_transaction(
-            transaction=tx,
-            sender=owner,
-            guardian=guardian_and_relayer_data.guardian,
-            relayer=guardian_and_relayer_data.relayer,
-            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
-            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
-        )
-
-        return tx
-
-    def prepare_transaction_for_automatic_activation(
-        self,
-        owner: IAccount,
-        delegation_contract: Address,
-        set: bool,
-        unset: bool,
-        gas_limit: int,
-        gas_price: int,
-        value: int,
-        nonce: int,
-        version: int,
-        options: int,
-        guardian_and_relayer_data: GuardianRelayerData,
-    ) -> Transaction:
-        if set and unset:
-            raise BadUsage("Cannot set and unset at the same time")
-
-        if set:
-            tx = self._factory.create_transaction_for_setting_automatic_activation(
-                sender=owner.address, delegation_contract=delegation_contract
-            )
-        elif unset:
-            tx = self._factory.create_transaction_for_unsetting_automatic_activation(
-                sender=owner.address, delegation_contract=delegation_contract
-            )
-        else:
-            raise BadUsage("Both set and unset automatic activation are False")
-
-        tx.value = value
-        tx.gas_price = gas_price
-        tx.nonce = nonce
-        tx.version = version
-        tx.options = options
-        tx.guardian = guardian_and_relayer_data.guardian_address
-        tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
-
-        self.sign_transaction(
-            transaction=tx,
-            sender=owner,
-            guardian=guardian_and_relayer_data.guardian,
-            relayer=guardian_and_relayer_data.relayer,
-            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
-            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
-        )
-
-        return tx
-
-    def prepare_transaction_for_redelegate_cap(
-        self,
-        owner: IAccount,
-        delegation_contract: Address,
-        set: bool,
-        unset: bool,
-        gas_limit: int,
-        gas_price: int,
-        value: int,
-        nonce: int,
-        version: int,
-        options: int,
-        guardian_and_relayer_data: GuardianRelayerData,
-    ) -> Transaction:
-        if set and unset:
-            raise BadUsage("Cannot set and unset at the same time")
-
-        if set:
-            tx = self._factory.create_transaction_for_setting_cap_check_on_redelegate_rewards(
-                sender=owner.address, delegation_contract=delegation_contract
-            )
-        elif unset:
-            tx = self._factory.create_transaction_for_unsetting_cap_check_on_redelegate_rewards(
-                sender=owner.address, delegation_contract=delegation_contract
-            )
-        else:
-            raise BadUsage("Either set or unset should be True")
-
-        tx.value = value
-        tx.gas_price = gas_price
-        tx.nonce = nonce
-        tx.version = version
-        tx.options = options
-        tx.guardian = guardian_and_relayer_data.guardian_address
-        tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
-
-        self.sign_transaction(
-            transaction=tx,
-            sender=owner,
-            guardian=guardian_and_relayer_data.guardian,
-            relayer=guardian_and_relayer_data.relayer,
-            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
-            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
-        )
-
-        return tx
-
-    def prepare_transaction_for_setting_metadata(
-        self,
-        owner: IAccount,
-        delegation_contract: Address,
-        name: str,
-        website: str,
-        identifier: str,
-        gas_limit: int,
-        gas_price: int,
-        value: int,
-        nonce: int,
-        version: int,
-        options: int,
-        guardian_and_relayer_data: GuardianRelayerData,
-    ) -> Transaction:
-        tx = self._factory.create_transaction_for_setting_metadata(
-            sender=owner.address,
-            delegation_contract=delegation_contract,
-            name=name,
-            website=website,
-            identifier=identifier,
-        )
-        tx.value = value
-        tx.gas_price = gas_price
-        tx.nonce = nonce
-        tx.version = version
-        tx.options = options
-        tx.guardian = guardian_and_relayer_data.guardian_address
-        tx.relayer = guardian_and_relayer_data.relayer_address
-
-        self.add_extra_gas_limit_if_required(tx)
-
-        if gas_limit:
-            tx.gas_limit = gas_limit
-
-        self.sign_transaction(
-            transaction=tx,
-            sender=owner,
-            guardian=guardian_and_relayer_data.guardian,
-            relayer=guardian_and_relayer_data.relayer,
-            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
-            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
-        )
-
-        return tx
-
-    # will be replaced in the future once it's implemented in sdk-py
-    def prepare_transaction_for_creating_delegation_contract_from_validator(
-        self,
-        owner: IAccount,
-        max_cap: int,
-        service_fee: int,
-        gas_limit: int,
-        gas_price: int,
-        value: int,
-        nonce: int,
-        version: int,
-        options: int,
-        guardian_and_relayer_data: GuardianRelayerData,
-    ) -> Transaction:
-        receiver = Address.new_from_hex(DELEGATION_MANAGER_SC_ADDRESS_HEX, get_address_hrp())
-
-        serializer = Serializer()
-        data = "makeNewContractFromValidatorData@" + serializer.serialize(
-            [BigUIntValue(max_cap), BigUIntValue(service_fee)]
-        )
-
-        tx = Transaction(
-            sender=owner.address,
+            contract=contract,
             receiver=receiver,
-            gas_limit=510000000,
-            chain_id=self._factory.config.chain_id,
-            data=data.encode(),
-            nonce=nonce,
-            version=version,
-            options=options,
-            guardian=guardian_and_relayer_data.guardian_address,
-            relayer=guardian_and_relayer_data.relayer_address,
-            gas_price=gas_price,
-            value=value,
+            native_token_amount=native_token_amount,
+            gas_limit=gas_limit,
+            opt_gas_limit=opt_gas_limit,
+            abi=abi,
+            function=function,
+            arguments=args,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
         )
 
-        if gas_limit:
-            tx.gas_limit = gas_limit
+        return tx
+
+    def prepare_transfer_execute_esdt_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        receiver: Address,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        should_prepare_args_for_factory: bool,
+        guardian_and_relayer_data: GuardianRelayerData,
+        token_transfers: list[TokenTransfer],
+        opt_gas_limit: Optional[int] = None,
+        abi: Optional[Abi] = None,
+        function: Optional[str] = None,
+        arguments: Optional[list[Any]] = None,
+    ) -> Transaction:
+        args = arguments if arguments else []
+        if should_prepare_args_for_factory:
+            args = self._convert_args_to_typed_values(args)
+
+        tx = self._factory.create_transaction_for_propose_transfer_esdt_execute(
+            sender=owner.address,
+            contract=contract,
+            receiver=receiver,
+            token_transfers=token_transfers,
+            gas_limit=gas_limit,
+            opt_gas_limit=opt_gas_limit,
+            abi=abi,
+            function=function,
+            arguments=args,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_async_call_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        receiver: Address,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        should_prepare_args_for_factory: bool,
+        guardian_and_relayer_data: GuardianRelayerData,
+        native_token_amount: int = 0,
+        token_transfers: Optional[list[TokenTransfer]] = None,
+        opt_gas_limit: Optional[int] = None,
+        abi: Optional[Abi] = None,
+        function: Optional[str] = None,
+        arguments: Optional[list[Any]] = None,
+    ) -> Transaction:
+        args = arguments if arguments else []
+        if should_prepare_args_for_factory:
+            args = self._convert_args_to_typed_values(args)
+
+        tx = self._factory.create_transaction_for_propose_async_call(
+            sender=owner.address,
+            contract=contract,
+            receiver=receiver,
+            gas_limit=gas_limit,
+            native_token_amount=native_token_amount,
+            token_transfers=token_transfers,
+            opt_gas_limit=opt_gas_limit,
+            abi=abi,
+            function=function,
+            arguments=args,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_contract_deploy_from_source_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        contract_to_copy: Address,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        upgradeable: bool,
+        readable: bool,
+        payable: bool,
+        payable_by_sc: bool,
+        should_prepare_args_for_factory: bool,
+        guardian_and_relayer_data: GuardianRelayerData,
+        native_token_amount: int = 0,
+        abi: Optional[Abi] = None,
+        arguments: Optional[list[Any]] = None,
+    ) -> Transaction:
+        args = arguments if arguments else []
+        if should_prepare_args_for_factory:
+            args = self._convert_args_to_typed_values(args)
+
+        tx = self._factory.create_transaction_for_propose_contract_deploy_from_source(
+            sender=owner.address,
+            contract=contract,
+            gas_limit=gas_limit,
+            contract_to_copy=contract_to_copy,
+            is_upgradeable=upgradeable,
+            is_readable=readable,
+            is_payable=payable,
+            is_payable_by_sc=payable_by_sc,
+            native_token_amount=native_token_amount,
+            abi=abi,
+            arguments=args,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_contract_upgrade_from_source_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        contract_to_upgrade: Address,
+        contract_to_copy: Address,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        upgradeable: bool,
+        readable: bool,
+        payable: bool,
+        payable_by_sc: bool,
+        should_prepare_args_for_factory: bool,
+        guardian_and_relayer_data: GuardianRelayerData,
+        native_token_amount: int = 0,
+        abi: Optional[Abi] = None,
+        arguments: Optional[list[Any]] = None,
+    ) -> Transaction:
+        args = arguments if arguments else []
+        if should_prepare_args_for_factory:
+            args = self._convert_args_to_typed_values(args)
+
+        tx = self._factory.create_transaction_for_propose_contract_upgrade_from_source(
+            sender=owner.address,
+            contract=contract,
+            contract_to_upgrade=contract_to_upgrade,
+            contract_to_copy=contract_to_copy,
+            gas_limit=gas_limit,
+            is_upgradeable=upgradeable,
+            is_readable=readable,
+            is_payable=payable,
+            is_payable_by_sc=payable_by_sc,
+            native_token_amount=native_token_amount,
+            abi=abi,
+            arguments=args,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_sign_action_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        action_id: int,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ) -> Transaction:
+        tx = self._factory.create_transaction_for_sign_action(
+            sender=owner.address,
+            contract=contract,
+            action_id=action_id,
+            gas_limit=gas_limit,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_sign_batch_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        batch_id: int,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ) -> Transaction:
+        tx = self._factory.create_transaction_for_sign_batch(
+            sender=owner.address,
+            contract=contract,
+            batch_id=batch_id,
+            gas_limit=gas_limit,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_sign_and_perform_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        action_id: int,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ) -> Transaction:
+        tx = self._factory.create_transaction_for_sign_and_perform(
+            sender=owner.address,
+            contract=contract,
+            action_id=action_id,
+            gas_limit=gas_limit,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_sign_batch_and_perform_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        batch_id: int,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ) -> Transaction:
+        tx = self._factory.create_transaction_for_sign_batch_and_perform(
+            sender=owner.address,
+            contract=contract,
+            batch_id=batch_id,
+            gas_limit=gas_limit,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_unsign_action_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        action_id: int,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ) -> Transaction:
+        tx = self._factory.create_transaction_for_unsign_action(
+            sender=owner.address,
+            contract=contract,
+            action_id=action_id,
+            gas_limit=gas_limit,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_unsign_batch_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        batch_id: int,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ) -> Transaction:
+        tx = self._factory.create_transaction_for_unsign_batch(
+            sender=owner.address,
+            contract=contract,
+            batch_id=batch_id,
+            gas_limit=gas_limit,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_unsign_for_outdated_board_members_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        action_id: int,
+        outdated_board_members: list[int],
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ) -> Transaction:
+        tx = self._factory.create_transaction_for_unsign_for_outdated_board_members(
+            sender=owner.address,
+            contract=contract,
+            action_id=action_id,
+            outdated_board_members=outdated_board_members,
+            gas_limit=gas_limit,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_perform_action_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        action_id: int,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ) -> Transaction:
+        tx = self._factory.create_transaction_for_perform_action(
+            sender=owner.address,
+            contract=contract,
+            action_id=action_id,
+            gas_limit=gas_limit,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
+
+        self.sign_transaction(
+            transaction=tx,
+            sender=owner,
+            guardian=guardian_and_relayer_data.guardian,
+            relayer=guardian_and_relayer_data.relayer,
+            guardian_service_url=guardian_and_relayer_data.guardian_service_url,
+            guardian_2fa_code=guardian_and_relayer_data.guardian_2fa_code,
+        )
+
+        return tx
+
+    def prepare_perform_batch_transaction(
+        self,
+        owner: IAccount,
+        nonce: int,
+        contract: Address,
+        batch_id: int,
+        gas_limit: int,
+        gas_price: int,
+        version: int,
+        options: int,
+        guardian_and_relayer_data: GuardianRelayerData,
+    ) -> Transaction:
+        tx = self._factory.create_transaction_for_perform_batch(
+            sender=owner.address,
+            contract=contract,
+            batch_id=batch_id,
+            gas_limit=gas_limit,
+        )
+        tx.nonce = nonce
+        tx.version = version
+        tx.options = options
+        tx.gas_price = gas_price
+        tx.guardian = guardian_and_relayer_data.guardian_address
+        tx.relayer = guardian_and_relayer_data.relayer_address
 
         self.sign_transaction(
             transaction=tx,
